requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
pandas==2.1.3
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
scikit-learn==1.3.2
spacy==3.7.2
keyring==24.3.0
cryptography==41.0.7
